requirements.txt
# Core dependencies
PyYAML>=6.0
cerberus>=1.3.4

# GUI dependencies (optional)
tkinter-tooltip>=2.0.0

# Media processing dependencies
Pillow>=9.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950

# Optional dependencies for advanced features
# ffmpeg-python>=0.2.0  # For proxy generation
# librosa>=0.9.0        # For audio analysis
# opencv-python>=4.5.0  # For thumbnail generation